Week 4
Set up RestKit correctly it seems, but still having an issue fetching
the XML, Professor said it was an issue with the headers so I set mine
to accept xml, but not difference.  Hopefully he’ll see what I missed
and correct.  Ignore the Programs class, was built to try a different
take on RestKit.